+++ src/mongo/mongo.py
@@ -13,6 +13,11 @@ async def setup():
     global client
     client = motor.motor_asyncio.AsyncIOMotorClient(os.getenv('MONGO_URL'))
 
+async def delete(db_name, collection_name, iden):
+    global client
+    collection = client[db_name][collection_name]
+    await collection.delete_one({"_id": iden})
+
 async def get_entire_collection(db_name, collection_name, return_one = False):
     global client
     collection = client[db_name][collection_name]